pyproject.toml
# pyproject.toml

[build-system]
requires = ["setuptools>=65.5.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "depman"
version = "0.1.0"
description = "CLI extension for Gitman: Wraps commands and checks for dependency updates"
readme = "README.md"
authors = [{ name = "Ivan MeL" }]
dependencies = ["click >= 8.0",
            "gitman>=0.11.0",
            "gitpython>=3.1.0","tqdm>=4.64.0",
            "rich>=13.3.0"]

[project.scripts]
depman = "depman.cli:cli"
